requirements.txt
pyperclip>=1.8.2
python-dotenv>=1.0.0
sv-ttk>=2.0.0
pystray>=0.19.5
Pillow>=10.0.0
